LayoutTests/media/modern-media-controls/volume-slider/volume-slider.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>VolumeSlider</code> class.");

const volumeSlider = new VolumeSlider;

shouldBeTrue('volumeSlider.element.classList.contains("volume")');

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
